refactor: move auth routes to module
